refactor(fixture): move fixtures to conftest
