+++ backend/calculatrice_monitoring/tests/test_eval.py
@@ -7,15 +7,6 @@ import pytest
 
 from calculatrice_monitoring.eval import evaluate
 
-from .fixtures import (
-    eval_context,
-    metadata,
-    monitoring_objects,
-    # FIXME: is there a way no to import those "unused" fixtures
-    protocols,
-    users,
-)
-
 
 class TestEvalIndicator:
     @pytest.mark.usefixtures("monitoring_objects")
